Configuration of environment variables was carried out to avoid uploading sensitive data to GitHub.
Reset Password by email.

User creation/validation by email.

The creation of goals was fixed in the Backend.

The cleaning of the inputs of the expense and goal income forms was configured after entry.

Correction of Warning due to ill-advised use of Redux when assigning the global variable users (FormLogin.js, line 27)